requirements.txt
requests
nltk
numpy
sentence-transformers
scikit-learn
transformers
python-dotenv
rapidfuzz
